Rename: controller -> presentation 패키지명 변경 및 RequestDto Update -> UpdateMyInfo
