+++ user-api/src/main/java/com/biengual/userapi/swagger/user/SwaggerUserUpdate.java
@@ -1,7 +1,7 @@
 package com.biengual.userapi.swagger.user;
 
 import com.biengual.userapi.swagger.SwaggerReturnInterface;
-import com.biengual.userapi.user.domain.dto.UserResponseDto;
+import com.biengual.userapi.user.presentation.UserResponseDto;
 
 public class SwaggerUserUpdate extends SwaggerReturnInterface<UserResponseDto.UserUpdateResponse> {
 }
