+++ user-api/src/main/java/com/biengual/userapi/user/presentation/UserRequestDto.java
@@ -1,4 +1,4 @@
-package com.biengual.userapi.user.domain.dto;
+package com.biengual.userapi.user.presentation;
 
 import java.time.LocalDate;
 
@@ -6,7 +6,7 @@ import com.biengual.userapi.user.domain.enums.Gender;
 
 public class UserRequestDto {
 
-	public record Update(
+	public record UpdateMyInfo(
 		String username,
 		String nickname,
 		String phoneNumber,
