+++ user-api/src/main/java/com/biengual/userapi/user/presentation/UserResponseDto.java
@@ -1,4 +1,4 @@
-package com.biengual.userapi.user.domain.dto;
+package com.biengual.userapi.user.presentation;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
